requirements.txt
transformers>=4.35
torch>=2.0
google-generativeai>=0.4.0
google-cloud-aiplatform>=1.43.0
docx
odfpy
gradio
pytest